The session progress can now be saved.
The session progress is serialised through the API calls. An
ordered-statistics tree is not saved, as it is just duplicating
the data to improve performance.

The user can select the origin/destination path using -p and -o calls
respectively. In the future, they may be specified with directories, so
there is space left for implementation.